pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opendart-llm-tools"
version = "0.1.0"
description = "Python tools for easily adding Korean DART filing retrieval to LLMs such as OpenAI and Gemini."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "openai>=1.0",
  "google-genai>=1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
